+++ Source/Urho3D/Graphics/OpenGL/OGLShaderProgram.cpp
@@ -20,6 +20,8 @@
 // THE SOFTWARE.
 //
 
+#include "../../Precompiled.h"
+
 #include "../../Graphics/ConstantBuffer.h"
 #include "../../Graphics/Graphics.h"
 #include "../../Graphics/GraphicsImpl.h"
@@ -65,7 +67,7 @@ ShaderProgram::~ShaderProgram()
 void ShaderProgram::OnDeviceLost()
 {
     GPUObject::OnDeviceLost();
-    
+
     if (graphics_ && graphics_->GetShaderProgram() == this)
         graphics_->SetShaders(0, 0);
 
@@ -78,19 +80,19 @@ void ShaderProgram::Release()
     {
         if (!graphics_)
             return;
-        
+
         if (!graphics_->IsDeviceLost())
         {
             if (graphics_->GetShaderProgram() == this)
                 graphics_->SetShaders(0, 0);
-            
+
             glDeleteProgram(object_);
         }
-        
+
         object_ = 0;
         linkerOutput_.Clear();
         shaderParameters_.Clear();
-        
+
         for (unsigned i = 0; i < MAX_TEXTURE_UNITS; ++i)
             useTextureUnit_[i] = false;
         for (unsigned i = 0; i < MAX_SHADER_PARAMETER_GROUPS; ++i)
@@ -101,17 +103,17 @@ void ShaderProgram::Release()
 bool ShaderProgram::Link()
 {
     Release();
-    
+
     if (!vertexShader_ || !pixelShader_ || !vertexShader_->GetGPUObject() || !pixelShader_->GetGPUObject())
         return false;
-    
+
     object_ = glCreateProgram();
     if (!object_)
     {
         linkerOutput_ = "Could not create shader program";
         return false;
     }
-    
+
     // Bind vertex attribute locations to ensure they are the same in all shaders
     // Note: this is not the same order as in VertexBuffer, instead a remapping is used to ensure everything except cube texture
     // coordinates fit to the first 8 for better GLES2 device compatibility
@@ -125,22 +127,22 @@ bool ShaderProgram::Link()
     glBindAttribLocation(object_, 7, "iBlendIndices");
     glBindAttribLocation(object_, 8, "iCubeTexCoord");
     glBindAttribLocation(object_, 9, "iCubeTexCoord2");
-    #ifndef GL_ES_VERSION_2_0
+#ifndef GL_ES_VERSION_2_0
     glBindAttribLocation(object_, 10, "iInstanceMatrix1");
     glBindAttribLocation(object_, 11, "iInstanceMatrix2");
     glBindAttribLocation(object_, 12, "iInstanceMatrix3");
-    #endif
-    
+#endif
+
     glAttachShader(object_, vertexShader_->GetGPUObject());
     glAttachShader(object_, pixelShader_->GetGPUObject());
     glLinkProgram(object_);
-    
+
     int linked, length;
     glGetProgramiv(object_, GL_LINK_STATUS, &linked);
     if (!linked)
     {
         glGetProgramiv(object_, GL_INFO_LOG_LENGTH, &length);
-        linkerOutput_.Resize(length);
+        linkerOutput_.Resize((unsigned)length);
         int outLength;
         glGetProgramInfoLog(object_, length, &outLength, &linkerOutput_[0]);
         glDeleteProgram(object_);
@@ -148,19 +150,19 @@ bool ShaderProgram::Link()
     }
     else
         linkerOutput_.Clear();
-    
+
     if (!object_)
         return false;
-    
+
     const int MAX_PARAMETER_NAME_LENGTH = 256;
     char uniformName[MAX_PARAMETER_NAME_LENGTH];
     int uniformCount;
-    
+
     glUseProgram(object_);
     glGetProgramiv(object_, GL_ACTIVE_UNIFORMS, &uniformCount);
-    
+
     // Check for constant buffers
-    #ifndef GL_ES_VERSION_2_0
+#ifndef GL_ES_VERSION_2_0
     HashMap<unsigned, unsigned> blockToBinding;
 
     if (Graphics::GetGL3Support())
@@ -171,9 +173,9 @@ bool ShaderProgram::Link()
         for (int i = 0; i < numUniformBlocks; ++i)
         {
             int nameLength;
-            glGetActiveUniformBlockName(object_, i, MAX_PARAMETER_NAME_LENGTH, &nameLength, uniformName);
+            glGetActiveUniformBlockName(object_, (GLuint)i, MAX_PARAMETER_NAME_LENGTH, &nameLength, uniformName);
 
-            String name(uniformName, nameLength);
+            String name(uniformName, (unsigned)nameLength);
 
             unsigned blockIndex = glGetUniformBlockIndex(object_, name.CString());
             unsigned group = M_MAX_UNSIGNED;
@@ -204,7 +206,7 @@ bool ShaderProgram::Link()
             if (group >= MAX_SHADER_PARAMETER_GROUPS)
             {
                 LOGWARNING("Skipping unrecognized uniform block " + name + " in shader program " + vertexShader_->GetFullName() +
-                    " " + pixelShader_->GetFullName());
+                           " " + pixelShader_->GetFullName());
                 continue;
             }
 
@@ -223,20 +225,20 @@ bool ShaderProgram::Link()
             glUniformBlockBinding(object_, blockIndex, bindingIndex);
             blockToBinding[blockIndex] = bindingIndex;
 
-            constantBuffers_[bindingIndex] = graphics_->GetOrCreateConstantBuffer(bindingIndex, dataSize);
+            constantBuffers_[bindingIndex] = graphics_->GetOrCreateConstantBuffer(bindingIndex, (unsigned)dataSize);
         }
     }
-    #endif
+#endif
 
     // Check for shader parameters and texture units
     for (int i = 0; i < uniformCount; ++i)
     {
         unsigned type;
         int count;
-        
-        glGetActiveUniform(object_, i, MAX_PARAMETER_NAME_LENGTH, 0, &count, &type, uniformName);
+
+        glGetActiveUniform(object_, (GLuint)i, MAX_PARAMETER_NAME_LENGTH, 0, &count, &type, uniformName);
         int location = glGetUniformLocation(object_, uniformName);
-        
+
         // Check for array index included in the name and strip it
         String name(uniformName);
         unsigned index = name.Find('[');
@@ -245,10 +247,10 @@ bool ShaderProgram::Link()
             // If not the first index, skip
             if (name.Find("[0]", index) == String::NPOS)
                 continue;
-            
+
             name = name.Substring(0, index);
         }
-        
+
         if (name[0] == 'c')
         {
             // Store constant uniform
@@ -257,7 +259,7 @@ bool ShaderProgram::Link()
             newParam.type_ = type;
             newParam.location_ = location;
 
-            #ifndef GL_ES_VERSION_2_0
+#ifndef GL_ES_VERSION_2_0
             // If running OpenGL 3, the uniform may be inside a constant buffer
             if (newParam.location_ < 0 && Graphics::GetGL3Support())
             {
@@ -270,7 +272,7 @@ bool ShaderProgram::Link()
                     newParam.bufferPtr_ = constantBuffers_[blockToBinding[blockIndex]];
                 }
             }
-            #endif
+#endif
 
             if (newParam.location_ >= 0)
                 shaderParameters_[StringHash(paramName)] = newParam;
@@ -291,7 +293,7 @@ bool ShaderProgram::Link()
                     }
                 }
             }
-            
+
             if (unit < MAX_TEXTURE_UNITS)
             {
                 useTextureUnit_[unit] = true;
@@ -299,10 +301,10 @@ bool ShaderProgram::Link()
             }
         }
     }
-    
+
     // Rehash the parameter map to ensure minimal load factor
     shaderParameters_.Rehash(NextPowerOfTwo(shaderParameters_.Size()));
-    
+
     return true;
 }
 
@@ -341,7 +343,7 @@ bool ShaderProgram::NeedParameterUpdate(ShaderParameterGroup group, const void* 
     }
 
     // The shader program may use a mixture of constant buffers and individual uniforms even in the same group
-    #ifndef GL_ES_VERSION_2_0
+#ifndef GL_ES_VERSION_2_0
     bool useBuffer = constantBuffers_[group].Get() || constantBuffers_[group + MAX_SHADER_PARAMETER_GROUPS].Get();
     bool useIndividual = !constantBuffers_[group].Get() || !constantBuffers_[group + MAX_SHADER_PARAMETER_GROUPS].Get();
     bool needUpdate = false;
@@ -359,7 +361,7 @@ bool ShaderProgram::NeedParameterUpdate(ShaderParameterGroup group, const void* 
     }
 
     return needUpdate;
-    #else
+#else
     if (parameterSources_[group] != source)
     {
         parameterSources_[group] = source;
@@ -367,13 +369,13 @@ bool ShaderProgram::NeedParameterUpdate(ShaderParameterGroup group, const void* 
     }
     else
         return false;
-    #endif
+#endif
 }
 
 void ShaderProgram::ClearParameterSource(ShaderParameterGroup group)
 {
     // The shader program may use a mixture of constant buffers and individual uniforms even in the same group
-    #ifndef GL_ES_VERSION_2_0
+#ifndef GL_ES_VERSION_2_0
     bool useBuffer = constantBuffers_[group].Get() || constantBuffers_[group + MAX_SHADER_PARAMETER_GROUPS].Get();
     bool useIndividual = !constantBuffers_[group].Get() || !constantBuffers_[group + MAX_SHADER_PARAMETER_GROUPS].Get();
 
@@ -381,9 +383,9 @@ void ShaderProgram::ClearParameterSource(ShaderParameterGroup group)
         globalParameterSources[group] = (const void*)M_MAX_UNSIGNED;
     if (useIndividual)
         parameterSources_[group] = (const void*)M_MAX_UNSIGNED;
-    #else
+#else
     parameterSources_[group] = (const void*)M_MAX_UNSIGNED;
-    #endif
+#endif
 }
 
 void ShaderProgram::ClearParameterSources()
@@ -392,10 +394,10 @@ void ShaderProgram::ClearParameterSources()
     if (!globalFrameNumber)
         ++globalFrameNumber;
 
-    #ifndef GL_ES_VERSION_2_0
+#ifndef GL_ES_VERSION_2_0
     for (unsigned i = 0; i < MAX_SHADER_PARAMETER_GROUPS; ++i)
         globalParameterSources[i] = (const void*)M_MAX_UNSIGNED;
-    #endif
+#endif
 }
 
 void ShaderProgram::ClearGlobalParameterSource(ShaderParameterGroup group)
